feat(getState): add ability to find state
Currently, this lib expects the state to be at the key `toastr` (and even handles when the state object is an immutable object (via `state.get`). There is no way to define a custom / different location easily.

This adds an optional prop to the `ReduxToastr` component which is a function that takes the current state to provide the toastr state.